move column, dfschema, etc. to common module
