+++ datafusion-common/Cargo.toml
@@ -42,3 +42,4 @@ parquet = { version = "8.0.0", features = ["arrow"] }
 avro-rs = { version = "0.13", features = ["snappy"], optional = true }
 pyo3 = { version = "0.15", optional = true }
 sqlparser = "0.13"
+ordered-float = "2.10"
