+++ datafusion-common/src/lib.rs
@@ -18,7 +18,13 @@
 mod column;
 mod dfschema;
 mod error;
+#[cfg(feature = "pyarrow")]
+mod pyarrow;
+mod scalar;
 
 pub use column::Column;
 pub use dfschema::{DFField, DFSchema, DFSchemaRef, ExprSchema, ToDFSchema};
 pub use error::{DataFusionError, Result};
+pub use scalar::{
+    ScalarType, ScalarValue, MAX_PRECISION_FOR_DECIMAL128, MAX_SCALE_FOR_DECIMAL128,
+};
